+++ crates/rpc/src/test_setup.rs
@@ -13,7 +13,7 @@ pub async fn test_storage<F: FnOnce(StateUpdate) -> StateUpdate>(
     version: StarknetVersion,
     customize_state_update: F,
 ) -> (Storage, BlockHeader, ContractAddress, ContractAddress) {
-    let storage = Storage::in_memory().unwrap();
+    let storage = pathfinder_storage::StorageBuilder::in_memory().unwrap();
     let mut db = storage.connection().unwrap();
     let tx = db.transaction().unwrap();