+++ crates/storage/src/connection/block.rs
@@ -482,7 +482,7 @@ mod tests {
 
     // Create test database filled with block headers.
     fn setup() -> (Connection, Vec<BlockHeader>) {
-        let storage = crate::Storage::in_memory().unwrap();
+        let storage = crate::StorageBuilder::in_memory().unwrap();
         let mut connection = storage.connection().unwrap();
         let tx = connection.transaction().unwrap();
 
@@ -658,7 +658,7 @@ mod tests {
 
         #[test]
         fn empty_chain_returns_none() {
-            let storage = crate::Storage::in_memory().unwrap();
+            let storage = crate::StorageBuilder::in_memory().unwrap();
             let mut db = storage.connection().unwrap();
             let db = db.transaction().unwrap();
 
@@ -691,7 +691,7 @@ mod tests {
 
         #[test]
         fn gap_in_chain() {
-            let storage = crate::Storage::in_memory().unwrap();
+            let storage = crate::StorageBuilder::in_memory().unwrap();
             let mut db = storage.connection().unwrap();
             let db = db.transaction().unwrap();
 
@@ -720,7 +720,7 @@ mod tests {
 
         #[test]
         fn empty_chain_returns_none() {
-            let storage = crate::Storage::in_memory().unwrap();
+            let storage = crate::StorageBuilder::in_memory().unwrap();
             let mut db = storage.connection().unwrap();
             let db = db.transaction().unwrap();
 
@@ -733,7 +733,7 @@ mod tests {
 
         #[test]
         fn target_without_parent_returns_target() {
-            let storage = crate::Storage::in_memory().unwrap();
+            let storage = crate::StorageBuilder::in_memory().unwrap();
             let mut db = storage.connection().unwrap();
             let db = db.transaction().unwrap();